CI: special commit content management, avoid re-doing the same commit

+++ scripts/New-PR.ps1
@@ -16,20 +16,38 @@ if (!$?) { throw "Error running git config: $LASTEXITCODE." }
 git config user.email $GitUserEmail
 if (!$?) { throw "Error running git config: $LASTEXITCODE." }
 
-Write-Output 'Generating and pushing a branch…'
+Write-Output 'Generating a branch…'
 git switch --force-create $BranchName
 if (!$?) { throw "Error running git switch: $LASTEXITCODE." }
 git commit --all --message $CommitMessage
 if (!$?) { throw "Error running git commit: $LASTEXITCODE." }
-git push --force --set-upstream origin $BranchName
-if (!$?) { throw "Error running git push: $LASTEXITCODE." }
 
 Write-Output 'Checking if a PR already exists…'
 [array] $issues = gh pr list --head $BranchName --json url | ConvertFrom-Json
 if (!$?) { throw "Error running gh pr list: $LASTEXITCODE." }
 if ($issues) {
     Write-Output "PR already exists: $($issues[0])."
+    Write-Output 'Comparing the local branch with the server one…'
+    $localTreeHash = git rev-parse 'HEAD^{tree}'
+    if (!$?) { throw "Error running git rev-parse: $LASTEXITCODE." }
+
+    $remoteTreeHash = git rev-parse "origin/$BranchName^{tree}"
+    if (!$?) { throw "Error running git rev-parse: $LASTEXITCODE." }
+
+    if ($localTreeHash -eq $remoteTreeHash) {
+        Write-Output "Local tree hash is the same as the remote tree hash: `"$localTreeHash`"."
+    } else {
+        Write-Output "Local tree hash `"$localTreeHash`" is not the same as the remote tree hash `"$remoteTreeHash`"."
+
+        Write-Output 'Force-pushing the branch…'
+        git push --force --set-upstream origin $BranchName
+        if (!$?) { throw "Error running git push: $LASTEXITCODE." }
+    }
 } else {
+    Write-Output 'Pushing the branch…'
+    git push --set-upstream origin $BranchName
+    if (!$?) { throw "Error running git push: $LASTEXITCODE." }
+
     Write-Output 'Creating a pull request…'
     gh pr create `
         --title $PrTitle `
